snakelr/pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snakelr"
version = "0.1.0"
description = "Gymnasium-compatible Snake environment with web and RLlib support."
readme = "README.md"
requires-python = ">=3.9"
authors = [
  {name="Your Name", email="you@example.com"}
]
dependencies = [
  "numpy>=1.24.0",
  "gymnasium>=0.29.1",
  "pygame>=2.5.2",
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "Pillow>=10.4.0",
  "ray[rllib]>=2.6.0",
]

[project.optional-dependencies]
web = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "Pillow>=10.4.0"
]
train = [
  "ray[rllib]>=2.6.0"
]

[tool.setuptools.package-dir]
"" = "."

[tool.setuptools.packages.find]
include = ["snakelr*"]
